requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
passlib==1.7.4
python-multipart==0.0.12
itsdangerous==2.2.0
requests==2.32.3
python-docx==1.1.2
reportlab==4.2.5
pytest==8.3.3
httpx==0.27.2
email-validator==2.2.0
chardet==5.2.0
